feat(analytics): refine spending insights and add test data utility
This commit introduces a centralized test data generator and refines the analytics logic to focus specifically on expense tracking. It also improves the home screen state by integrating recent transaction history.

- **`TestDataGenerator.kt`**:
    - Created a new singleton utility to generate randomized budgets and transactions.
    - Implemented logic to ensure expenses are represented as negative values to align with application logic.
- **`AnalyticsViewModel.kt` & `AnalyticsUiRoute.kt`**:
    - Updated trend, heatmap, and habit insight calculations to filter for expenses (`amount < 0`).
    - Applied `.absoluteValue` to spending totals for clearer visualization in trend charts.
    - Adjusted habit insight thresholds and set the habits section to be expanded by default.
- **`HomeViewModel.kt`**:
    - Updated `uiState` to use `combine`, merging financial profile data with the transaction repository.
    - Now populates the success state with the 10 most recent transactions.
    - Delegated random transaction generation to `TestDataGenerator`.
- **`SettingsViewModel.kt`**:
    - Refactored to use `TestDataGenerator`, removing redundant local data generation logic.
- **`CategoryGridRoute.kt`**:
    - Standardized Hilt ViewModel imports and organized UI component references.
