+++ applications/seaweed/apps/mobile/features/home/src/main/java/com/zoewave/probase/seaweed/mobile/home/ui/HomeViewModel.kt
@@ -4,11 +4,11 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.zoewave.probase.seaweed.data.FinancialRepository
 import com.zoewave.probase.seaweed.data.TransactionRepository
-import com.zoewave.probase.seaweed.model.Transaction
+import com.zoewave.probase.seaweed.data.TestDataGenerator
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import java.util.UUID
@@ -17,21 +17,24 @@ import javax.inject.Inject
 @HiltViewModel
 class HomeViewModel @Inject constructor(
     private val repository: TransactionRepository,
-    private val financialRepository: FinancialRepository
+    private val financialRepository: FinancialRepository,
+    private val testDataGenerator: TestDataGenerator
 ) : ViewModel() {
 
-    val uiState: StateFlow<HomeUiState> = financialRepository.getFinancialProfile()
-        .map { profile ->
-            HomeUiState.Success(
-                transactions = emptyList(), // We might need recent transactions here
-                categoriesSummary = profile.categoryOverviews,
-                monthlyIncome = profile.monthlyIncome,
-                totalFixedCosts = profile.totalFixedCosts,
-                flexibleMoneyRemaining = profile.flexibleMoneyRemaining,
-                unallocatedMoney = profile.unallocatedMoney,
-                monthProgress = profile.monthProgress
-            )
-        }.stateIn(
+    val uiState: StateFlow<HomeUiState> = combine(
+        financialRepository.getFinancialProfile(),
+        repository.getAllTransactions()
+    ) { profile, transactions ->
+        HomeUiState.Success(
+            transactions = transactions.sortedByDescending { it.date }.take(10),
+            categoriesSummary = profile.categoryOverviews,
+            monthlyIncome = profile.monthlyIncome,
+            totalFixedCosts = profile.totalFixedCosts,
+            flexibleMoneyRemaining = profile.flexibleMoneyRemaining,
+            unallocatedMoney = profile.unallocatedMoney,
+            monthProgress = profile.monthProgress
+        )
+    }.stateIn(
             scope = viewModelScope,
             started = SharingStarted.WhileSubscribed(5_000),
             initialValue = HomeUiState.Loading
@@ -46,15 +49,7 @@ class HomeViewModel @Inject constructor(
             }
             HomeUiEvent.AddRandomTransaction -> {
                 viewModelScope.launch {
-                    val categories = listOf("Food", "Transport", "Rent", "Entertainment", "Salary", "Investment")
-                    val randomTransaction = Transaction(
-                        id = UUID.randomUUID().toString(),
-                        amount = (10..10000).random().toDouble() / 100.0 * (if ((0..1).random() == 0) 1 else -1),
-                        category = categories.random(),
-                        description = "Random transaction",
-                        date = System.currentTimeMillis() - (0..30L * 24 * 60 * 60 * 1000).random()
-                    )
-                    repository.addTransaction(randomTransaction)
+                    testDataGenerator.generateSingleRandomTransaction()
                 }
             }
             HomeUiEvent.OnBackClicked -> { /* Handled in Route */ }
