+++ applications/seaweed/apps/mobile/features/settings/src/main/java/com/zoewave/probase/seaweed/mobile/settings/ui/SettingsViewModel.kt
@@ -2,8 +2,7 @@ package com.zoewave.probase.seaweed.mobile.settings.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.zoewave.probase.seaweed.data.BudgetTargetRepository
-import com.zoewave.probase.seaweed.data.TransactionRepository
+import com.zoewave.probase.seaweed.data.TestDataGenerator
 import com.zoewave.probase.seaweed.data.UserSettingsRepository
 import com.zoewave.probase.seaweed.model.BudgetTarget
 import com.zoewave.probase.seaweed.model.SeaweedThemeConfig
@@ -20,8 +19,7 @@ import kotlin.random.Random
 @HiltViewModel
 class SettingsViewModel @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
-    private val transactionRepository: TransactionRepository,
-    private val budgetRepository: BudgetTargetRepository,
+    private val testDataGenerator: TestDataGenerator,
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow<SettingsUiState>(SettingsUiState.Loading)
@@ -49,55 +47,10 @@ class SettingsViewModel @Inject constructor(
                     userSettingsRepository.saveUserSettings(currentSettings.copy(monthlyIncome = event.income))
                 }
                 is SettingsUiEvent.GenerateTestData -> {
-                    generateRandomTransactions()
+                    testDataGenerator.generateThreeMonthsOfData()
                 }
                 is SettingsUiEvent.NavigateTo -> { /* Handled in Route */ }
             }
-        }
-    }
-
-    private suspend fun generateRandomTransactions() {
-        val categories = listOf("Food", "Transport", "Shopping", "Entertainment", "Health", "Utilities")
-        
-        // Generate random budgets first
-        categories.forEach { category ->
-            val limit = when (category) {
-                "Food" -> Random.nextDouble(300.0, 600.0)
-                "Transport" -> Random.nextDouble(100.0, 300.0)
-                "Shopping" -> Random.nextDouble(200.0, 800.0)
-                "Entertainment" -> Random.nextDouble(100.0, 400.0)
-                "Health" -> Random.nextDouble(50.0, 200.0)
-                "Utilities" -> Random.nextDouble(200.0, 500.0)
-                else -> 500.0
-            }
-            budgetRepository.saveBudget(BudgetTarget(category, limit))
-        }
-
-        val now = System.currentTimeMillis()
-        val ninetyDaysMillis = 90L * 24 * 60 * 60 * 1000
-
-        repeat(150) {
-            val randomTimeOffset = Random.nextLong(0, ninetyDaysMillis)
-            val date = now - randomTimeOffset
-            val category = categories.random()
-            val amount = when (category) {
-                "Food" -> Random.nextDouble(5.0, 50.0)
-                "Transport" -> Random.nextDouble(2.0, 30.0)
-                "Shopping" -> Random.nextDouble(10.0, 200.0)
-                "Entertainment" -> Random.nextDouble(10.0, 100.0)
-                "Health" -> Random.nextDouble(20.0, 150.0)
-                "Utilities" -> Random.nextDouble(50.0, 300.0)
-                else -> Random.nextDouble(1.0, 100.0)
-            }
-
-            val transaction = Transaction(
-                id = UUID.randomUUID().toString(),
-                amount = amount,
-                category = category,
-                description = "Random $category expense",
-                date = date
-            )
-            transactionRepository.addTransaction(transaction)
         }
     }
 }
